칵테일 좋아요 상위 10개 구현 중

+++ backend/src/main/java/com/ssafy/cocktail/backend/cocktails/dto/CocktailSortedLikes.java
@@ -4,10 +4,10 @@ import lombok.Getter;
 
 @Getter
 public class CocktailSortedLikes {
-    private int cocktailId;
+    private Long cocktailId;
     private String cocktailNameKo;
     private String cocktailNameEn;
     private String cocktailImg;
     private Double cocktailRating;
-    private int cnt;
+    private Long cnt;
 }
